requirements.txt
django
psycopg2-binary
Pillow
django-silk
django-ninja
djangorestframework-simplejwt